backend/pyproject.toml
[project]
name = "digital-signage-backend"
version = "1.0.0"
requires-python = ">=3.12"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0",
  "asyncpg",
  "alembic>=1.13",
  "pydantic>=2.6",
  "pydantic-settings",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "bcrypt==4.0.1",
  "redis>=5.0",
  "python-multipart",
  "uvicorn[standard]",
  "psycopg[binary]",
  "pydantic[email]",
  "passlib[bcrypt]>=1.7.4"
]
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
